refactor(heatmaps): move heatmap-subpipeline.ts from analytics to heatmaps dir
The heatmap subpipeline is now only used by the heatmaps consumer
(after the joined pipeline lost its branching). Move the file to
heatmaps/ alongside the consumer and pipeline. Drop the now-stale
re-export from analytics/index.ts.

Behavior preserved 1:1 — processGroupsStep stays so that group writes
happen on heatmap events the same way they do on master.
