+++ nodejs/src/ingestion/analytics/index.ts
@@ -1,7 +1,5 @@
 export { createEventSubpipeline, EventSubpipelineConfig, EventSubpipelineInput } from './event-subpipeline'
 
-export { createHeatmapSubpipeline, HeatmapSubpipelineConfig, HeatmapSubpipelineInput } from './heatmap-subpipeline'
-
 export {
     createPerDistinctIdPipeline,
     PerDistinctIdPipelineConfig,
